Protect metadata and stop traffic analysis; added DKIM sig; fixed several bugs

+++ goodcrypto/mail/message/bundle.py
@@ -1,6 +1,6 @@
 '''
     Copyright 2014-2015 GoodCrypto
-    Last modified: 2015-07-27
+    Last modified: 2015-11-28
 
     This file is open source, licensed under GPLv3 <http://www.gnu.org/licenses/>.
 '''
@@ -20,10 +20,11 @@ from goodcrypto.mail.internal_settings import get_domain
 from goodcrypto.mail.message import constants, history, utils
 from goodcrypto.mail.message.crypto_message import CryptoMessage
 from goodcrypto.mail.message.email_message import EmailMessage
-from goodcrypto.mail.message.encrypt_utils import create_protected_message
-from goodcrypto.mail.utils import get_metadata_address, get_sysadmin_email
+from goodcrypto.mail.message.encrypt_utils import add_dkim_sig_optionally, create_protected_message
+from goodcrypto.mail.message.metadata import get_metadata_address, parse_bundled_message
+from goodcrypto.mail.utils import get_encryption_software, get_sysadmin_email, send_message
 from goodcrypto.mail.utils.dirs import get_packet_directory, SafeDirPermissions
-from goodcrypto.mail.utils.notices import notify_user
+from goodcrypto.mail.utils.notices import report_bad_bundled_encrypted_message
 from goodcrypto.oce.crypto_factory import CryptoFactory
 from goodcrypto.oce.open_pgp_analyzer import OpenPGPAnalyzer
 from goodcrypto.utils import i18n, get_email
@@ -32,10 +33,10 @@ from goodcrypto.utils.log_file import LogFile
 from syr import mime_constants
 
 class Bundle(object):
-    ''' 
+    '''
         Bundle and pad messages to each domain we have their metadata address.
-        
-        Unlike much of GoodCrypto, the results of many functions in this class are 
+
+        Unlike much of GoodCrypto, the results of many functions in this class are
         python style email.Messages, not EmailMessages or CryptoMessages.
         Comments try to use camel back to make the disticion clear (i.e., Message
         referrs to an email.Message().
@@ -50,13 +51,13 @@ class Bundle(object):
         self.DEBUGGING = False
 
         self.log = None
-        
+
         self.bundled_messages = []
         self.crypted_with = []
-        
+
     def bundle_and_pad(self):
         ''' Bundle and pad messages to reduce tracking. '''
-    
+
         packet_dir = get_packet_directory()
         dirnames = os.listdir(packet_dir)
         if dirnames is None or len(dirnames) <= 0:
@@ -89,33 +90,33 @@ class Bundle(object):
                 sender = get_email(get_metadata_address(domain=get_domain()))
                 recipient = get_email(get_metadata_address(domain=to_domain))
                 self.log_message('starting to send message from {} to {}'.format(sender, recipient))
-                result_ok = utils.send_message(sender, recipient, message.as_string())
+                result_ok = send_message(sender, recipient, message.as_string())
                 self.log_message('finished sending message')
         except Exception as exception:
             result_ok = False
             self.log_message('error while sending message')
             self.log_message('EXCEPTION - see goodcrypto.utils.exception.log for details')
             record_exception()
-    
+
         return result_ok
-    
+
     def create_message(self, dirname, to_domain):
         ''' Create a Message to send that contains any other messages that are ready. '''
-    
+
         inner_message = self.create_inner_message(dirname, to_domain)
         if inner_message is None:
             encrypted_message = None
             self.log_message('no inner message for {}'.format(to_domain))
         else:
             encrypted_message = self.create_encrypted_message(inner_message, to_domain)
-    
+
         return encrypted_message
-    
+
     def create_encrypted_message(self, inner_message, to_domain):
         ''' Create an encrypted Message. '''
-    
+
         message = None
-    
+
         if to_domain is None:
             self.log_message('domain is not defined')
         elif inner_message is None:
@@ -123,12 +124,16 @@ class Bundle(object):
         else:
             from_user = get_email(get_metadata_address(domain=get_domain()))
             to_user = get_email(get_metadata_address(domain=to_domain))
-            encryption_names = utils.get_encryption_software(to_user)
+            encryption_names = get_encryption_software(to_user)
 
             crypto_message = create_protected_message(
                 from_user, to_user, inner_message.as_string(), utils.get_message_id())
 
             if crypto_message.is_crypted():
+
+                # add the DKIM signature to the inner message if user opted for it
+                crypto_message = add_dkim_sig_optionally(crypto_message)
+
                 message = crypto_message.get_email_message().get_message()
                 self.crypted_with = crypto_message.is_crypted_with()
                 for part in message.walk():
@@ -136,15 +141,15 @@ class Bundle(object):
                     if self.DEBUGGING:
                         self.log_message(part.get_payload())
             else:
-                self.report_bad_encrypted_message(to_domain)
-            
+                report_bad_bundled_encrypted_message(to_domain, self.bundled_messages)
+
         return message
 
     def create_inner_message(self, dirname, to_domain):
         ''' Create a Message that contains other messages plus padding. '''
 
         message = None
-    
+
         if dirname is None:
             self.log_message('no dir name defined in create_inner_message')
         elif to_domain is None:
@@ -182,9 +187,9 @@ class Bundle(object):
                         self.log_message(part.get_payload())
             else:
                 self.log_message('Unable to get any parts so no inner message created')
-            
+
         return message
-    
+
     def pad_message(self, parts, to_domain):
         ''' Pad the Message with random characters. '''
 
@@ -192,7 +197,7 @@ class Bundle(object):
         original_size = 0
         for part in parts:
             original_size += len(part.as_string())
-    
+
         # then pad the message with random characters
         current_size = original_size
         target_size = options.bundled_message_max_size()
@@ -207,32 +212,32 @@ class Bundle(object):
               base64.b64encode(rnd_bytes), mime_constants.ALTERNATIVE_SUB_TYPE, encode_base64)
             current_size += len(part.as_string())
             parts.append(part)
-            
+
         self.log_message('padded with {} random bytes'.format(target_size - original_size))
-    
+
         return parts
-        
+
     def get_mime_part(self, dirname, filename, estimated_size):
-        ''' Get a MIME part with the Message. 
-    
+        ''' Get a MIME part with the Message.
+
             If any errors occur, then bounce the message to the original sender.
         '''
         part = None
-    
+
         fullname = os.path.join(dirname, filename)
         filesize = os.stat(fullname).st_size
         max_size = options.bundled_message_max_size()
-    
+
         # only look at the message files that won't make the overall size too large
-        if (filename.startswith(constants.MESSAGE_PREFIX) and 
+        if (filename.startswith(constants.MESSAGE_PREFIX) and
             filename.endswith(constants.MESSAGE_SUFFIX)):
-    
+
             # if the message is too large, then bounce it back to the user
             if filesize > max_size:
-                self.bounce_message(fullname, 
+                self.bounce_message(fullname,
                   i18n('Message too large to send. It must be {size} KB or smaller'.format(
                          size=options.bundle_message_kb())))
-    
+
             elif (filesize + estimated_size) < max_size:
                 try:
                     with open(fullname) as f:
@@ -247,15 +252,15 @@ class Bundle(object):
                 except:
                     record_exception()
                     self.log_message('EXCEPTION - see goodcrypto.utils.exception.log for details')
-    
+
             else:
                 self.log_message('{} is too large for this batch of messages (estimated size with this message: {})'.format(filename, filesize + estimated_size))
-    
+
         return part, filesize
-    
+
     def prep_message_header(self, message, to_domain):
         ''' Prepare the header of a Message. '''
-    
+
         if message is None:
             self.log_message('no message defined in prep_message_header')
         elif to_domain is None:
@@ -264,7 +269,7 @@ class Bundle(object):
             message_date = datetime.utcnow().replace(tzinfo=utc)
             from_user = get_metadata_address(domain=get_domain())
             to_user = get_metadata_address(domain=to_domain)
-        
+
             message.__setitem__(mime_constants.FROM_KEYWORD, from_user)
             message.__setitem__(mime_constants.TO_KEYWORD, to_user)
             message.__setitem__(constants.ORIGINAL_FROM, from_user)
@@ -293,7 +298,7 @@ class Bundle(object):
         if len(self.bundled_messages) > 0:
             for bundled_message in self.bundled_messages:
                 with open(bundled_message) as f:
-                    original_message, addendum = utils.parse_bundled_message(f.read())
+                    original_message, addendum = parse_bundled_message(f.read())
                     encrypted = get_addendum_value(addendum, constants.CRYPTED_KEYWORD)
                     if encrypted:
                         sender = get_addendum_value(addendum, mime_constants.FROM_KEYWORD)
@@ -310,20 +315,23 @@ class Bundle(object):
                         self.log_message('logged headers in goodcrypto.message.utils.log')
                         utils.log_message_headers(crypto_message, tag='bundled headers')
                         self.log_message('added encrypted history record from {}'.format(sender))
-                        
-                os.remove(bundled_message)
+
+                if os.path.exists(bundled_message):
+                    os.remove(bundled_message)
+                else:
+                    self.log_message('tried to delete message after bundling it, but message no longer exists on disk')
         else:
             self.log_message('no bundled messages')
-        
+
     def bounce_message(self, fullname, error_message):
         ''' Bounce a Message to the original user. '''
-        
+
         notified_user = False
-        
+
         if fullname is not None and os.path.exists(fullname):
             with open(fullname) as f:
                 content = f.read()
-                original_message, addendum = utils.parse_bundled_message(content)
+                original_message, addendum = parse_bundled_message(content)
                 subject = i18n('{} - Unable to send message to {email}'.format(
                     TAG_ERROR, email=addendum[mime_constants.TO_KEYWORD]))
                 notified_user = utils.bounce_message(
@@ -331,27 +339,6 @@ class Bundle(object):
                 self.log_message(subject)
 
         return notified_user
-    
-    def report_bad_encrypted_message(self, to_domain):
-        ''' Report unable to create an encrypted bundled message. '''
-        
-        subject = i18n('{} - Unable to send messages to {domain}'.format(TAG_WARNING, domain=to_domain))
-
-        line1 = i18n('Your GoodCrypto private server tried to send messages to {domain} using the _no_metadata_ keys. It was unable to do so.'.format(domain=to_domain))
-        line2 = i18n("You should verify that you have a contact and key for both your domain and {domain}'s domain.".format(domain=to_domain))
-        line3 = i18n("You can disable bundling and padding messages, but it means that your users will be easier to track.")
-
-        # leave a trailing space in case we add a 4th line
-        sysadmin_message = '{}\n\n{}\n\n{} '.format(line1, line2, line3)
-        
-        if len(self.bundled_messages) > 0:
-            line4 = i18n("Also, {} messages to {domain} will be lost if you disable bundling before resolving the current problem.".format(
-                len(self.bundled_messages), domain=to_domain))
-            sysadmin_message += line4
-
-        sysadmin = get_sysadmin_email()
-        notify_user(sysadmin, subject, sysadmin_message)
-        self.log_message('sent notice to {}\n{}'.format(sysadmin, sysadmin_message))
 
     def log_message(self, message):
         '''
@@ -359,6 +346,6 @@ class Bundle(object):
         '''
         if self.log is None:
             self.log = LogFile()
-    
+
         self.log.write_and_flush(message)
